refactor: 기존 Delegate패턴과 ViewController 타입 모두 삭제 후 리팩토링 #7
- 출력을 분리하여 만들어주었던 ViewController와 그에 따른 Delegate 프로토콜을 모두 삭제
- 이후 Delegate 프로퍼티 삭제와 main 내부 코드 전반적으로 수정
