+++ BankManagerConsoleApp/BankManagerConsoleApp/Model/BankClerk.swift
@@ -2,15 +2,14 @@ import Foundation
 
 final class BankClerk {
     private let workSpeed: TimeInterval
-    var delegate: BankClerkDelegate?
     
     init(workSpeed: TimeInterval = 0.7) {
         self.workSpeed = workSpeed
     }
     
     func work(for customer: Customer) {
-        delegate?.bankClerk(didStartWork: customer.waitingNumber)
+        print("\(customer.waitingNumber)번 고객 업무 시작")
         Thread.sleep(forTimeInterval: workSpeed)
-        delegate?.bankClerk(didFinishWork: customer.waitingNumber)
+        print("\(customer.waitingNumber)번 고객 업무 완료")
     }
 }
