+++ BankManagerConsoleApp/BankManagerConsoleApp/Model/BankManager.swift
@@ -7,27 +7,39 @@ enum Menu {
 
 struct BankManager {
     private let bank: Bank
-    var delegate: BankManagerDelegate?
     
     init(bank: Bank) {
         self.bank = bank
     }
     
-    mutating func receivedUserInput(_ input: String) {
+    mutating func receiveUserInput() -> Bool {
+        guard let input = readLine() else {
+            showInvalidInput()
+            return true
+        }
         switch input {
         case Menu.open:
             setUpBankCustomers()
             bank.openBank()
+            return true
         case Menu.close:
-            delegate?.bankManagerDidClose()
-            return
+            return false
         default:
-            delegate?.bankManagerDidReceiveInvalidInput()
+            showInvalidInput()
+            return true
         }
     }
     
     private func setUpBankCustomers(to range: ClosedRange<Int> = 10...30) {
         let numberOfCustomers = Int.random(in: range)
         bank.handOutWaitingNumber(from: numberOfCustomers)
+    }
+    
+    func showMenu() {
+        print(BankManagerMessage.menuList, terminator: "")
+    }
+    
+    private func showInvalidInput() {
+        print(BankManagerMessage.wrongInput)
     }
 }
