--- BankManagerConsoleApp/BankManagerConsoleApp/Protocol/BankClerkDelegate.swift
@@ -1,6 +1,0 @@
-import Foundation
-
-protocol BankClerkDelegate {
-    func bankClerk(didStartWork customerNumber: Int)
-    func bankClerk(didFinishWork customerNumber: Int)
-}
