--- BankManagerConsoleApp/BankManagerConsoleApp/Protocol/BankDelegate.swift
@@ -1,5 +1,0 @@
-import Foundation
-
-protocol BankDelegate {
-    func bank(didFinish numberOfCustomers: Int, durationTime: String)
-}
