--- BankManagerConsoleApp/BankManagerConsoleApp/Protocol/BankManagerDelegate.swift
@@ -1,6 +1,0 @@
-import Foundation
-
-protocol BankManagerDelegate {
-    func bankManagerDidReceiveInvalidInput()
-    func bankManagerDidClose()
-}
