+++ BankManagerConsoleApp/BankManagerConsoleApp/main.swift
@@ -1,15 +1,13 @@
 import Foundation
 
 func run() {
-    let bankClerk = BankClerk()
+    let bankClerk = BankClerk(workSpeed: 0.7)
     let bank = Bank(bankClerk: bankClerk)
-    let bankManager = BankManager(bank: bank)
-    let _ = BankClerkViewController(bankClerk: bankClerk)
-    let _ = BankViewController(bank: bank)
-    let bankManagerViewController = BankManagerViewController(bankManager: bankManager)
-    while bankManagerViewController.isRunning {
-        bankManagerViewController.showMenu()
-        bankManagerViewController.receiveUserInput()
+    var bankManager = BankManager(bank: bank)
+    var isRunning = true
+    while isRunning {
+        bankManager.showMenu()
+        isRunning = bankManager.receiveUserInput()
     }
 }
